pyproject.toml
[project]
name = "doc-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.22.0",
    "pre-commit>=4.5.1",
    "tqdm>=4.67.1",
]
